# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitness AI API Documentation",
        "description": "L5 Swagger OpenApi description for the Fitness Backend",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/v1/gamification/leaderboard/weekly": {
            "get": {
                "tags": [
                    "Gamification"
                ],
                "summary": "Ranking semanal",
                "description": "Retorna o ranking de XP da semana atual.",
                "operationId": "b9d7e27f6cd59c293625c5161db054df",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Quantidade de resultados (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ranking semanal"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/gamification/leaderboard/monthly": {
            "get": {
                "tags": [
                    "Gamification"
                ],
                "summary": "Ranking mensal",
                "description": "Retorna o ranking de XP do mês atual.",
                "operationId": "6805c05fe982c246a6bf1c6c1869d141",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Quantidade de resultados (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ranking mensal"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/gamification/leaderboard/alltime": {
            "get": {
                "tags": [
                    "Gamification"
                ],
                "summary": "Ranking geral",
                "description": "Retorna o ranking de XP de todos os tempos.",
                "operationId": "6a7e2e159adf69c01c447c06e0145473",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Quantidade de resultados (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ranking geral"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/gamification/leaderboard/friends": {
            "get": {
                "tags": [
                    "Gamification"
                ],
                "summary": "Ranking de amigos",
                "description": "Retorna o ranking de XP entre amigos aceitos e o usuário autenticado.",
                "operationId": "0b1314dedcf6f84460b2ae69a84388ca",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Quantidade de resultados (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ranking de amigos"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/meals/analyze-text": {
            "post": {
                "tags": [
                    "Meals"
                ],
                "summary": "Analisa e salva uma refeição a partir de texto",
                "description": "O usuário envia o que comeu (em texto) e a IA estima macros, calorias e salva no histórico.",
                "operationId": "4c378a1e3c612f05586da43c5e23ba58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-15"
                                    },
                                    "meal_type": {
                                        "type": "string",
                                        "example": "lunch"
                                    },
                                    "text_description": {
                                        "type": "string",
                                        "example": "Comi 150g de frango, 200g de arroz e salada de alface."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Refeição salva com sucesso"
                    }
                }
            }
        },
        "/api/v1/meals/analyze-image": {
            "post": {
                "tags": [
                    "Meals"
                ],
                "summary": "Analisa e salva uma refeição a partir de uma FOTO",
                "description": "O usuário envia a foto do prato em base64 e a IA usa visão computacional para estimar macros, calorias e itens.",
                "operationId": "8252733e771ab550889746bad7d9be34",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-15"
                                    },
                                    "meal_type": {
                                        "type": "string",
                                        "example": "dinner"
                                    },
                                    "image_base64": {
                                        "type": "string",
                                        "example": "base64_string_here"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Refeição computada e salva via foto"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}